+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - add initial setup for IRM params: `ri` and `Tcrit`
 
+### Fixed
+- fix `maxBorrow` estimation
+
 ## [0.15.0] - 2024-12-02
 ### Added
 - `PublicAllocator` contract for vaults